Creating observation reader file and example nb
Adding utility for reading in the radiosonde data as a
dictionary of xarrays. I'm not sure if there is a better
way to read them in without creating several 'time'
dimensions to combine them into one xarray dataset. In
the end, that may be the way to go, but for now they are
a dictionary of xarray datasets.

The example plotting is only for observations at the
moment. We may add examples of plotting with WRF / FE
output.